task-9-async-chatrooms/requirements.txt
gunicorn>=19.9.0
aiohttp>=3.5.4
aiohttp-jinja2>=1.1.0
jinja2>=2.10.1
aiohttp-session>=2.7.0
aioredis>=1.2.0
aiopg>=0.16.0
pyyaml>=5.1
sqlalchemy>=1.3.3
faker>=1.0.5
aiohttp_swagger>=1.0.5
pylint>=2.2
pytest>=4.4.0
pytest-cov>=2.6
pytest-aiohttp>=0.3.0
pytest-mock>=1.10
flake8>=3.7
black>=18.9b0